Display button to accelerate on markets page
