+++ client/webserver/site/src/css/icons.scss
@@ -56,6 +56,10 @@
   content: "\e902";
 }
 
+.ico-rocket::before {
+  content: "\e9a5";
+}
+
 .ico-profile::before {
   content: "\e903";
 }
