+++ client/webserver/site/src/css/market.scss
@@ -696,6 +696,11 @@ div.numorders {
     font-size: 12px;
     cursor: pointer;
   }
+
+  .ico-rocket {
+    font-size: 12px;
+    cursor: pointer;
+  }
 }
 
 #vDetailPane {
