+++ client/webserver/site/src/html/order.tmpl
@@ -84,11 +84,13 @@
         <div>[[[Age]]]</div>
         <div data-stamp="{{$ord.Stamp}}"></div>
       </div>
-      <div class="order-datum d-hide" id="accelerateDiv">
-        <div>[[[accelerate]]]</div>
-        <div><button id="accelerateBttn">[[[accelerate]]]</button></div>
-      </div>
     </div> {{- /* END DATA CARDS */ -}}
+
+    {{- /* ACTIONS */ -}}
+    <div class="fs18 demi pt-4 pb-2 d-hide" id="actionsLabel">[[[Actions]]]</div>
+    <div class="d-flex align-items-stretch justify-content-start flex-wrap">
+      <button id="accelerateBttn" type="button" class="px-2 py-1 fs15 bg2 selected d-hide">[[[accelerate]]]</button>
+    </div>
 
     {{- /* MATCHES */ -}}
     <div class="fs18 demi pt-4 pb-2{{if eq (len $ord.Matches) 0}} d-hide{{end}}" id="matchesLabel">[[[Matches]]]</div>
@@ -223,79 +225,9 @@
       {{template "cancelOrderForm" .}}
     </form>
     <form class="d-hide" id="accelerateForm"  autocomplete="off">
-      {{$passwordIsCached := .UserInfo.PasswordIsCached}}
-      <div class="form-closer hoverbg"><span class="ico-cross"></span></div>
-      <div class="text-center fs22 sans-light position-relative">
-        [[[accelerate_order]]]
-      </div>
-      <div id=accelerateMainDiv>
-        <div id="configureAccelerationDiv">
-          [[[acceleration_text]]]
-          <br />
-          <div class="text-center fs17 sans-light position-relative">
-            [[[effective_swap_tx_rate]]]: <span id="accelerateAvgFeeRate"></span>
-          </div>
-          <div class="text-center fs17 sans-light position-relative">
-            [[[current_fee]]]: <span id="accelerateCurrentFeeRate"></span>
-          </div>
-          <div id="sliderContainer" class="slider-container">
-          </div>
-          <div id="feeEstimateDiv" class="text-center fs17 sans-light position-relative d-hide">
-            [[[acceleration_cost_msg]]]
-          </div>
-          <hr class="dashed mt-2">
-          <div class="d-flex flex-row align-items-end {{if $passwordIsCached}}justify-content-end{{end}} pb-4">
-            <div class="col-12 p-0 {{if $passwordIsCached}}d-hide{{end}}">
-              <label for="acceleratePass" class="pt-3 mb-1">[[[Password]]]</label>
-              <input type="password" class="form-control select" id="acceleratePass" autocomplete="off">
-            </div>
-            <div class="col-12 pt-2 {{if not $passwordIsCached}}px-5{{end}}">
-              <button id="accelerateSubmit" type="button" class="w-100 fs15 bg2 selected">[[[Submit]]]</button>
-            </div>
-          </div>
-        </div>
-        <div class="d-hide" id="earlyAccelerationDiv">
-          <div id="recentAccelerationMsg">[[[recent_acceleration_msg]]]</div>
-          <div id="recentSwapMsg">[[[recent_swap_msg]]]</div>
-          <br/>
-          [[[early_acceleration_help_msg]]]
-          <div class="d-flex flex-row align-items-end justify-content-end pb-4">
-            <div class="col-12 pt-2">
-              <button id="submitEarlyConfirm" type="button" class="w-100 fs15 bg2 selected">[[[Submit]]]</button>
-            </div>
-          </div>
-        </div>
-        <div class="fs15 pt-3 text-center d-hide errcolor text-break" id="accelerateErr"></div>
-      </div>
-      <div id="accelerateMsgDiv" class="d-hide">
-        <div class="fs15 pt-3 text-center d-hide errcolor text-break" id="preAccelerateErr"></div>
-        <div class="fs15 pt-3 text-left d-hide text-break" id="accelerateSuccess">
-          [[[accelerate_success]]]
-        </div>
-      </div>
+      {{template "accelerateForm" .}}
     </form> 
   </div>
-
-  <div id="rangeOptTmpl">
-    <div class="d-flex flex-nowrap fs12">
-      <span data-tmpl="rangeLblStart"></span>
-      <div data-tmpl="slider" class="slider flex-grow-1">
-        <div data-tmpl="handle" class="slider-handle"></div>
-      </div>
-      <span data-tmpl="rangeLblEnd"></span>
-    </div>
-    <div class="d-flex justify-content-between align-items-center fs14">
-      <span>
-        <input type="text" pattern="[0-9]+([\.,][0-9]+)?" data-tmpl="xInput" class="xy-range-input d-hide">
-        <span data-tmpl="x" class="pointer"></span> <span data-tmpl="xUnit"></span>
-      </span>
-      <span>
-        <input type="text" pattern="[0-9]+([\.,][0-9]+)?" data-tmpl="yInput" class="xy-range-input d-hide">
-        <span data-tmpl="y" class="pointer"></span> <span data-tmpl="yUnit"></span>
-      </span>
-    </div>
-  </div>
-
 </div>
 {{template "bottom"}}
 {{end}}
